+++ generators/project/_package.js
@@ -10,7 +10,8 @@ module.exports = function () {
         },
 
         typescript: {
-            'gulp-typescript': '^2.13.6'
+            'gulp-typescript': '^2.13.6',
+            typings: "^1.0.5"
         }
     }
 
@@ -34,18 +35,43 @@ module.exports = function () {
 
     const getFramework = {
         angularjs: {
-            'gulp-ng-annotate': '^2.0.0',
+            'angular': '^1.5.6',
+            'gulp-ng-annotate': '^2.0.0'
         },
 
-        jquery: {},
+        angular2: {
+            '@angular/common': '2.0.0-rc.1',
+            '@angular/compiler': '2.0.0-rc.1',
+            '@angular/core': '2.0.0-rc.1',
+            '@angular/http': '2.0.0-rc.1',
+            '@angular/platform-browser': '2.0.0-rc.1',
+            '@angular/platform-browser-dynamic': '2.0.0-rc.1',
+            '@angular/router': '2.0.0-rc.1',
+            '@angular/router-deprecated': '2.0.0-rc.1',
+            '@angular/upgrade': '2.0.0-rc.1',
+            'angular2-in-memory-web-api': '0.0.11',
+            'bootstrap': '^3.3.6',
+            'core-js': '^2.4.0',
+            'reflect-metadata': '^0.1.3',
+            'rxjs': '5.0.0-beta.6',
+            'systemjs': '0.19.27',
+            'zone.js': '^0.6.12'
+        },
+
+        jquery: {
+            'jquery': '^3.0.0'
+        },
 
         polymer: {
             'gulp-size': '^2.1.0',
             'gulp-vulcanize': '^6.1.0',
+            'Polymer': '^1.4.0'
         },
 
         react: {
             'gulp-react': '^3.1.0',
+            'react': '^15.1.0',
+            'react-dom': '^15.1.0'
         },
 
         vanillajs: {}
@@ -79,22 +105,22 @@ module.exports = function () {
             'phantomjs-prebuilt': '^2.1.7'
         }
     }
+
     const file = (fs.existsSync('./package.json') === true)
         ? JSON.parse(fs.readFileSync('./package.json', 'utf8'))
         : {}
 
     const packDefault = {
-        'description': `${this.config.get('project')} with ${this.config.get('framework')}`,
-        'name': `${this.config.get('project')}`,
-        'private': true,
-        'version': '0.0.0',
-
+        description: `${this.config.get('project')} with ${this.config.get('framework')}`,
+        name: `${this.config.get('project')}`,
+        private: true,
+        version: '0.0.0',
         dependencies: {
             'babel-preset-es2015': '^6.6.0',
             'browser-sync': '^2.12.8',
             'del': '^2.2.0',
-            'express': '^4.13.1',
             'extend': '^3.0.0',
+            'http-proxy-middleware': '^0.15.2',
             'gulp': '^3.9.0',
             'gulp-babel': '^6.1.2',
             'gulp-changed': '^1.2.1',
@@ -116,21 +142,25 @@ module.exports = function () {
             'gulp-watch': '^4.3.5',
             'lodash': '^4.13.1',
             'plato': '^1.5.0',
-            'request': '^2.58.0',
             'run-sequence': '^1.1.1'
         },
 
-        'scripts': {
-            'ss-build': 'gulp build --open=false',
-            'ss-build-open': 'gulp build',
-            'ss-clean': 'gulp clean-dirs',
-            'ss-component': 'yo speedseed:component',
-            'ss-dist': 'gulp build --dist=true --server=false',
-            'ss-dist-open': 'gulp build --dist=true',
-            'ss-indent': 'gulp indent',
-            'ss-reports': 'gulp reports',
-            'ss-test': 'gulp test',
-            'ss-update': 'npm i generator-speedseed -g    &&    yo speedseed    &&    npm i'
+        engines : {
+            node : ">=4.0.0"
+        },
+
+        scripts: {
+            'build': 'gulp build --open=false',
+            'build-open': 'gulp build',
+            'clean': 'gulp clean-dirs',
+            'component': 'yo speedseed:component',
+            'dist': 'gulp build --dist=true --server=false',
+            'dist-open': 'gulp build --dist=true',
+            'indent': 'gulp indent',
+            'spec': 'gulp test',
+            'reports': 'gulp reports',
+            'typings': 'typings install',
+            'update': 'npm i generator-speedseed -g    &&    yo speedseed    &&    npm i'
         }
     }
 
@@ -151,5 +181,5 @@ module.exports = function () {
         file
     )
 
-    fs.writeFileSync('package.json', JSON.stringify(packDefault, null, 4))
+    fs.writeFileSync('package.json', JSON.stringify(packDefault, null, 2))
 }