+++ generators/seed/template/multi-tic-tac-toe/angular2/compiler/typescript/app/js/main.ts
@@ -1,0 +1,4 @@
+import { bootstrap } from '@angular/platform-browser-dynamic'
+import { SSTable } from './../components/table/main'
+
+bootstrap(SSTable)