+++ generators/seed/template/multi-tic-tac-toe/angularjs/compiler/typescript/app/components/box/main.ts
@@ -1,3 +1,5 @@
+/// <reference path="../../../typings/index.d.ts" />
+
 angular
 .module('speedseed')
 .component('ssBox', {