+++ generators/seed/template/multi-tic-tac-toe/polymer/compiler/typescript/component/main.ts
@@ -1,3 +1,5 @@
+/// <reference path="../../../typings/index.d.ts" />
+
 (() => {
 Polymer({
     is: '<%= component %>',