+++ generators/seed/template/multi-tic-tac-toe/react/compiler/babeljs/component/main.jsx
@@ -8,7 +8,7 @@
         }
     })
 
-    React.render(
+    ReactDOM.render(
         ({%= include(__dirname, '-dom.html') %}),
         document.getElementsByTagName('<%= component %>')[0]
     )