+++ generators/seed/template/multi-tic-tac-toe/vanillajs/compiler/typescript/component/main.ts
@@ -1,3 +1,5 @@
+/// <reference path="../../../typings/index.d.ts" />
+
 (($component) => {
     for (let i = 0, l = $component.length; i < l; i++) {
         $component[i].addEventListener('mousedown', function (e) {